Image displaying and category listing
During this stage of development, I went on to optimise the csv file reader, I created a concept logo for the site and began developing the table for the page. I created a function that loaded the referenced images and put in a placeholder if they could not be found. This was done for the search results / search page. I also split up the categories that were found in each row and made them more readable.