Product editor tool development and early stages of creating forms
During this new stage of development, I went on to develop the editor.html and /editor webpage for the proposed product management tool. The early stages of the form saw me run into some difficulties, as I wanted to create a reactive dropdown bootstrap menu that allowed immediate responses to user requests. The way I wanted the page to function was the user would select an ID from a dropdown list, and the product's respective properties would load into the fields.