Reactive dropdown menu with javascript
When I continued development of the product editor form I wanted to make the product selection process as easy as possbile. To do this I wanted the user to be able to select an ID from a list of IDs and then proceed to have all the data load up right away without having to send a post request. Hence, I used javascript to create an onchange algorithm that detected when an ID was selected and then sent the relevant information from Flask to the html. However, this meant the data had to be sent as a JSON string and then parsed into an object. The algorithm can be found in the documentation.