HTML form development and reactive Javascript
In the latest addition to the website, development of the reactive form was finally finished, with the user being able to request data for any product and then replace it's product information. Javascript was used to display the product data and once edited, the results where sent to app.py. These results at this stage were doing nothing, as the server-side python for the Post request had not be created yet.