Optimisations and Form Handling
For the editor html page, I began testing whether user changes to the product form worked correctly by testing whether the server received the data. This was successful so I began to work towards creating an algorithm for replacing the old product data with the new data and then writing it to the file.