Fixed None error with search function and attempted file writing with editor
In the editor function in the python file I began trying to get the data submitted by the user and writing it directly to the data file. I did however run into several problems, as I had not considered that the user may have chosen to not input any data at all creating a 'None type error'. When I came across this error it reminded me off the category and search selection function I had created earlier, where I did indeed discover that it suffered from the same error. This led to me not only fixing this issue but also optimising several areas of the code I created earlier in an attempt to improve both its overall perforamce and readability.