Completed editor function
After troubleshooting several issues related to file encoding and reading/writing the file, I was able to get the editor function working. The code for this section was also rewritten, as I was now able to anticipate what steps I would have to undergo to create the algorithm. I also added error checking to certain form values that may return as empty or none. The code worked as follows:
Fetched and checked the form data
Organised it into an array to be pushed to the file
Create a temporary array containing

CSV lines
Find line with data to replace
Replace line in temporary array
Open CSV file in write format
Write temporary array with new data to CSV

I also added the search bar to the editor page, so that the user can access the same search functions from the editor page.