Optimised product editing and comment implementation
Began implementing comments throughout the code so that the code is reusable and readable for future coding additions.

Added a render function for the new home page template, and added several redirects to this page in case the user accidently types in a different link.

In the search function no major changes were made, but some variable names were modified to ensure readability.

In the editor function I removed the now defunct checking if certain fields existed, as the new editor form ensures that all fields have to be filled out before submission.