Form validation and minor optimisations
In the javascript element of this template I added an if statement that checked if the product was currently in any existing categories. If not, it provided a message that notified of this.

I also added an error message in the form of a p tag at the top of the page that let the user know if their last form submission was successful.

Validation and input checking was also enabled on the form to ensure that all fields were filled out before being sent to the Flask serverside.