Created two home pages
Two home pages were developed by building of two different bootstrap example pages found on their website. The 'old homepage' was messy and did not follow the standard layout that I was aiming to achieve.

However, the newer homepage was a lot cleaner and provided the user with useful insight as to the purpose of each webpage.