Colour scheme implementation
This is hopefully the final major update for the html templates. In this update I applied the colour scheme to the templates to increase their relevance and useability for GGG. I also added the reset data button to the editor.html, which linked with its new function on the python serverside.